requirements.txt
cycler==0.11.0
fonttools==4.28.3
kiwisolver==1.3.2
matplotlib==3.5.1
mysql-connector-python==8.0.27
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyparsing==3.0.6
python-dateutil==2.8.2
scipy==1.7.3
six==1.16.0
